lkjadsf

+++ main.js
@@ -11069,14 +11069,14 @@ let setWorkout = function(self) {
     slideNext: function() {
       if(this.activeIndex !== this.images) {
         let value = 100 / this.images;
-        this.transform = parseInt(this.transform) + value;
+        this.transform = parseFloat(this.transform) + value;
         this.activeIndex++;
       }
     },
     slidePrev: function() {
       if(this.activeIndex !== 1) {
         let value = 100 / this.images;
-        this.transform = parseInt(this.transform) - value;
+        this.transform = parseFloat(this.transform) - value;
         this.activeIndex--;
       }
     }
